requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic-settings==2.6.0
pydantic[email]==2.10.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
aiosqlite==0.20.0
redis==5.2.0
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.0
langdetect==1.0.9
hijri-converter==2.3.0
pytz==2024.2
structlog==24.4.0httpx>=0.28.0
slowapi>=0.1.9
slowapi==0.1.9
